+++ lil_pwny/__main__.py
@@ -1,3 +1,0 @@
-from lil_pwny import main
-
-main()